Moved interface
moved interface from GroupResponse to interfaces and deleted GroupResponse.
