--- src/Group/GroupResponse.tsx
@@ -1,8 +1,0 @@
-export default interface IGroupResponse {
-    _id : string,
-    name : string,
-    maxSize : number,
-    game : string,
-    invite_id : string,
-    users : string[]
-};
