+++ src/services/interfaces.tsx
@@ -8,3 +8,12 @@ export interface UserService {
     setUserInfo(userId : string, groupId : string) : void;
     getUserInfo() : User;
 }
+
+export default interface IGroupResponse {
+    _id : string,
+    name : string,
+    maxSize : number,
+    game : string,
+    invite_id : string,
+    users : string[]
+};
